fix viewer threading, try a test result
remove copy of output

try test-results

test-reports in correct location

better test attempt

+++ Cktgen/flow.sh
@@ -9,7 +9,6 @@ OUTPUTVOL=outputVol
 ROUTERVOL=routerStrawman
 SKIPROUTER=NO
 SKIPGENERATE=NO
-STARTVIEWER=NO
 SHOWGLOBALROUTES=""
 SHOWMETALTEMPLATES=""
 ROUTE=" --route"
@@ -119,6 +118,7 @@ echo PLACERJSON = "${PLACERJSON}"
 echo SOURCE = "${SOURCE}"
 echo SMALL = "${SMALL}"
 
+
 M_INPUT="--mount source=${INPUTVOL},target=/Cktgen/INPUT"
 M_INPUT_VIEWER="--mount source=${INPUTVOL},target=/public/INPUT"
 M_out="--mount source=${OUTPUTVOL},target=/Cktgen/out"
@@ -128,9 +128,6 @@ docker volume rm -f ${ROUTERVOL}
 (cd ${TECHDIR} && tar cvf - .) | docker run --rm ${M_DR_COLLATERAL} -i ubuntu bash -c "cd /Cktgen/DR_COLLATERAL && tar xvf -"
 
 if [ ${SKIPGENERATE} = "NO" ]; then
-    if [ ${STARTVIEWER} = "NO" ]; then
-	docker volume rm -f ${INPUTVOL}
-    fi
     docker volume rm -f ${OUTPUTVOL}
     docker run --rm ${M_INPUT} ${M_DR_COLLATERAL} cktgen bash -c "source /general/bin/activate && cd /Cktgen && python ${SCRIPT} -n mydesign ${ROUTE}${SHOWGLOBALROUTES}${SHOWMETALTEMPLATES}${SOURCE}${PLACERJSON}${SMALL}"
 fi
@@ -141,6 +138,6 @@ if [ ${SKIPROUTER} = "NO" ]; then
     docker run --rm ${M_out} ${M_INPUT} ${M_DR_COLLATERAL} cktgen bash -c "source /general/bin/activate; cd /Cktgen && python ${SCRIPT} --consume_results -n mydesign ${SOURCE}${PLACERJSON}${SMALL}"
 fi
 
-if [ ${STARTVIEWER} = "YES" ]; then
+if [[ -v STARTVIEWER ]] && [[ ${STARTVIEWER} = "YES" ]]; then
     docker run --rm ${M_INPUT_VIEWER} -p${PORT}:8000 -d viewer_image /bin/bash -c "source /sympy/bin/activate && cd /public && python -m http.server"
 fi
